refactor: move server utils to separate file
